+++ controller/aigateway/onpremconfig/zz_generated.controller_aigatewaydataplanecertificate.go
@@ -46,11 +46,11 @@ import (
 )
 
 // -----------------------------------------------------------------------------
-// aiconfigurationv1alpha1 AIGatewayModel - Reconciler
+// aiconfigurationv1alpha1 AIGatewayDataPlaneCertificate - Reconciler
 // -----------------------------------------------------------------------------
 
-// AIGatewayModelReconciler reconciles AIGatewayModel resources
-type AIGatewayModelReconciler struct {
+// AIGatewayDataPlaneCertificateReconciler reconciles AIGatewayDataPlaneCertificate resources
+type AIGatewayDataPlaneCertificateReconciler struct {
 	client.Client
 
 	Log              logr.Logger
@@ -62,13 +62,13 @@ type AIGatewayModelReconciler struct {
 	Cache            map[types.NamespacedName]types.NamespacedName
 }
 
-var _ controllers.Reconciler = &AIGatewayModelReconciler{}
+var _ controllers.Reconciler = &AIGatewayDataPlaneCertificateReconciler{}
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *AIGatewayModelReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *AIGatewayDataPlaneCertificateReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	blder := ctrl.NewControllerManagedBy(mgr).
 		// set the controller name
-		Named("aiconfigurationv1alpha1AIGatewayModel").
+		Named("aiconfigurationv1alpha1AIGatewayDataPlaneCertificate").
 		WithOptions(controller.Options{
 			LogConstructor: func(_ *reconcile.Request) logr.Logger {
 				return r.Log
@@ -82,7 +82,7 @@ func (r *AIGatewayModelReconciler) SetupWithManager(mgr ctrl.Manager) error {
 				r.StatusQueue.Subscribe(schema.GroupVersionKind{
 					Group:   "aiconfiguration.konghq.com",
 					Version: "v1alpha1",
-					Kind:    "AIGatewayModel",
+					Kind:    "AIGatewayDataPlaneCertificate",
 				}),
 				&handler.EnqueueRequestForObject{},
 			),
@@ -90,7 +90,7 @@ func (r *AIGatewayModelReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	}
 	return blder.
 		For(
-			&aiconfigurationv1alpha1.AIGatewayModel{},
+			&aiconfigurationv1alpha1.AIGatewayDataPlaneCertificate{},
 			builder.WithPredicates(
 				GenerationChangedPredicate{},
 			),
@@ -99,12 +99,12 @@ func (r *AIGatewayModelReconciler) SetupWithManager(mgr ctrl.Manager) error {
 }
 
 // SetLogger sets the logger.
-func (r *AIGatewayModelReconciler) SetLogger(l logr.Logger) {
+func (r *AIGatewayDataPlaneCertificateReconciler) SetLogger(l logr.Logger) {
 	r.Log = l
 }
 
 // SetCommonFields sets the shared controller dependencies.
-func (r *AIGatewayModelReconciler) SetCommonFields(
+func (r *AIGatewayDataPlaneCertificateReconciler) SetCommonFields(
 	client client.Client,
 	scheme *runtime.Scheme,
 	log logr.Logger,
@@ -118,17 +118,17 @@ func (r *AIGatewayModelReconciler) SetCommonFields(
 	r.ChangeNotifier = changeNotifier
 }
 
-//+kubebuilder:rbac:groups=aiconfiguration.konghq.com,resources=aigatewaymodels,verbs=get;list;watch
-//+kubebuilder:rbac:groups=aiconfiguration.konghq.com,resources=aigatewaymodels/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=aiconfiguration.konghq.com,resources=aigatewaydataplanecertificates,verbs=get;list;watch
+//+kubebuilder:rbac:groups=aiconfiguration.konghq.com,resources=aigatewaydataplanecertificates/status,verbs=get;update;patch
 
 // Reconcile processes the watched objects
-func (r *AIGatewayModelReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+func (r *AIGatewayDataPlaneCertificateReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	if r.Cache == nil {
 		r.Cache = make(map[types.NamespacedName]types.NamespacedName)
 	}
-	obj := new(aiconfigurationv1alpha1.AIGatewayModel)
+	obj := new(aiconfigurationv1alpha1.AIGatewayDataPlaneCertificate)
 	logger := r.Log.
-		WithValues("aiconfigurationv1alpha1AIGatewayModel", req.NamespacedName).
+		WithValues("aiconfigurationv1alpha1AIGatewayDataPlaneCertificate", req.NamespacedName).
 		WithValues("namespace", req.Namespace).
 		WithValues("name", req.Name).
 		V(logging.DebugLevel.Value())
@@ -142,7 +142,7 @@ func (r *AIGatewayModelReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 			if parent, ok := r.Cache[req.NamespacedName]; ok {
 				obj.Namespace = req.Namespace
 				obj.Name = req.Name
-				obj.Kind = "AIGatewayModel"
+				obj.Kind = "AIGatewayDataPlaneCertificate"
 				if r.ChangeNotifier != nil {
 					r.ChangeNotifier.NotifyChange(ctx, &parent, obj)
 				}